+++ my-project/backend/ledger_api/app/api/services/report/concrete_generators.py
@@ -1,0 +1,51 @@
+# backend/app/api/services/concrete_generators.py
+
+import pandas as pd
+from .base_report_generator import BaseReportGenerator
+
+from app.api.st_app.logic.manage.factory_report import process as process_factory_report
+from app.api.st_app.logic.manage.balance_sheet import process as process_balance_sheet
+from app.api.st_app.logic.manage.management_sheet import (
+    process as process_management_sheet,
+)
+
+
+class FactoryReportGenerator(BaseReportGenerator):
+    """工場レポート生成クラス"""
+
+    def _main_process_impl(self) -> pd.DataFrame:
+        result_df = process_factory_report(self.files)
+        return result_df
+
+
+class BalanceSheetGenerator(BaseReportGenerator):
+    """バランスシート生成クラス"""
+
+    def _main_process_impl(self) -> pd.DataFrame:
+        result_df = process_balance_sheet(self.files)
+        return result_df
+
+
+class AverageSheetGenerator(BaseReportGenerator):
+    """平均シート生成クラス"""
+
+    def _main_process_impl(self) -> pd.DataFrame:
+        result_df = process_average_sheet(self.files)
+        return result_df
+
+
+class BlockUnitPriceGenerator(BaseReportGenerator):
+    """ブロック単価生成クラス"""
+
+    def _main_process_impl(self) -> pd.DataFrame:
+        # result_df = process_average_sheet(self.files)
+        # return result_df
+        return
+
+
+class ManagementSheetGenerator(BaseReportGenerator):
+    """管理シート生成クラス"""
+
+    def _main_process_impl(self) -> pd.DataFrame:
+        result_df = process_management_sheet(self.files)
+        return result_df
