+++ my-project/backend/ledger_api/app/local_config/paths.py
@@ -2,3 +2,6 @@ import os
 
 MANAGE_REPORT_OUTPUT_DIR = "/backend/app/api/endpoints/static/manage_report"
 MANAGE_REPORT_URL_BASE = "/ledger_api/static/manage_report"
+DEBUG_MANAGE_REPORTDIR = "/backend/app/data/csv/debug/manage_report"
+
+BASE_ST_APP_DIR = "/backend/app/api/st_app"
